Overload the create method for the admin interface to avoid any session specific info being required
Since we are using a different auth system for the admin we can't reuse all the code for the admin
interface, and we shouldn't so using the overloading provided by ActiveAdmin we override the defaults
at the controller level for petition and can create them directly from the admin dashboard.

The same will need to be done for most models. And that's a good thing.