Made a custom editor for constants based off the work Drew did
The work Drew did was a good base, but with the new system needed to
change it a bit.

ActiveAdmin didn't like the constants so I had to make my own editor for
them, tested it to work. We need to add some security to the update now, since the
admin login is different. Opening issue for that.